+++ common/src/main/java/com/binaris/wizardry/core/mixin/client/MouseHandlerMixin.java
@@ -11,12 +11,11 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 @Mixin(MouseHandler.class)
 public abstract class MouseHandlerMixin {
 
-    @Inject(method = "onScroll", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/player/LocalPlayer;isSpectator()Z"), locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true)
+    @Inject(method = "onScroll", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/player/LocalPlayer;isSpectator()Z"), cancellable = true)
     public void EBWIZARDRY$onScroll(long windowPointer, double xOffset, double yOffset, CallbackInfo ci) {
         Player player = Minecraft.getInstance().player;
         ItemStack wand = EntityUtil.getWandInUse(player);